+++ crates/ryo3-std/src/net/ipaddr.rs
@@ -2,7 +2,7 @@
 use crate::net::{PySocketAddrV4, PySocketAddrV6};
 use pyo3::prelude::*;
 use pyo3::types::PyTuple;
-use ryo3_macro_rules::err_py_not_impl;
+use ryo3_macro_rules::pytodo;
 use std::net::{IpAddr, Ipv4Addr, Ipv6Addr, SocketAddrV4, SocketAddrV6};
 
 #[cfg_attr(feature = "serde", derive(serde::Serialize))]
@@ -196,7 +196,7 @@ impl PyIpv4Addr {
     #[expect(clippy::unused_self)]
     #[getter]
     fn is_benchmarking(&self) -> PyResult<bool> {
-        err_py_not_impl!()
+        pytodo!()
     }
 
     #[getter]
@@ -212,7 +212,7 @@ impl PyIpv4Addr {
     #[expect(clippy::unused_self)]
     #[getter]
     fn is_global(&self) -> PyResult<bool> {
-        err_py_not_impl!()
+        pytodo!()
     }
 
     #[getter]
@@ -238,13 +238,13 @@ impl PyIpv4Addr {
     #[expect(clippy::unused_self)]
     #[getter]
     fn is_reserved(&self) -> PyResult<bool> {
-        err_py_not_impl!()
+        pytodo!()
     }
 
     #[expect(clippy::unused_self)]
     #[getter]
     fn is_shared(&self) -> PyResult<bool> {
-        err_py_not_impl!()
+        pytodo!()
     }
 
     #[getter]
@@ -367,7 +367,7 @@ impl PyIpv6Addr {
     #[getter]
     #[expect(clippy::unused_self)]
     fn is_benchmarking(&self) -> PyResult<bool> {
-        err_py_not_impl!()
+        pytodo!()
     }
 
     #[getter]
@@ -381,7 +381,7 @@ impl PyIpv6Addr {
     #[getter]
     #[expect(clippy::unused_self)]
     fn is_global(&self) -> PyResult<bool> {
-        err_py_not_impl!()
+        pytodo!()
     }
 
     #[getter]
@@ -410,7 +410,7 @@ impl PyIpv6Addr {
     #[getter]
     #[expect(clippy::unused_self)]
     fn is_unicast_global(&self) -> PyResult<bool> {
-        err_py_not_impl!()
+        pytodo!()
     }
 
     #[getter]
@@ -574,7 +574,7 @@ impl PyIpAddr {
     #[expect(clippy::unused_self)]
     #[getter]
     fn is_benchmarking(&self) -> PyResult<bool> {
-        err_py_not_impl!()
+        pytodo!()
     }
 
     #[getter]
@@ -634,7 +634,7 @@ impl PyIpAddr {
     #[getter]
     #[expect(clippy::unused_self)]
     fn is_global(&self) -> PyResult<bool> {
-        err_py_not_impl!()
+        pytodo!()
     }
 
     #[getter]
@@ -661,13 +661,13 @@ impl PyIpAddr {
     #[getter]
     #[expect(clippy::unused_self)]
     fn is_reserved(&self) -> PyResult<bool> {
-        err_py_not_impl!()
+        pytodo!()
     }
 
     #[getter]
     #[expect(clippy::unused_self)]
     fn is_shared(&self) -> PyResult<bool> {
-        err_py_not_impl!()
+        pytodo!()
     }
 
     #[getter]
@@ -681,7 +681,7 @@ impl PyIpAddr {
     #[getter]
     #[expect(clippy::unused_self)]
     fn is_unicast_global(&self) -> PyResult<bool> {
-        err_py_not_impl!()
+        pytodo!()
     }
 
     #[getter]